Stop using Result to indicate stopped searches
This means the return value from all search methods is just Eval (so
just a wrapped i32).

We now set an explicit 'stopped' flag which we can use to check if the
search was aborted, since we no longer get this information via the
return type.

This simplifies a lot of the search methods which had to wrap scores in
Ok().

I suspect it's also a minor speedup.

Elo   | 1.80 +- 3.53 (95%)
SPRT  | 8.0+0.08s Threads=1 Hash=16MB
LLR   | 2.95 (-2.94, 2.94) [-5.00, 0.00]
Games | N: 12766 W: 3388 L: 3322 D: 6056
Penta | [173, 1448, 3086, 1492, 184]