datagen: Fix bug that caused incorrect WDL records
WDL, prior to this change, was treated inconsistently. Since it wasn't
clear from the name whether it signified a player-relative win, or a
white win, different parts of the code took different approaches.

For example, game_result returned results relative to the current player
and these were corrected later on with code to interpret the result
based on the current player - as intended. But adjudicate_result
returned results from White's perspective, which was adjusted by the
same code. If the current player was white, nothing would change, but if
the current player was black, the result would be incorrectly flipped.

bench: 20406643