Do NNUE accumulator updates lazily
If we store the accumulators inside of Game and update them on every
move, there will be times we never end up using the values.

Instead, we can store the updates we need to make, and only make them
when we actually evaluate the position.

To do this efficiently, we store a stack of accumulators (where index 0
is the position being searched). Each successive entry in the stack
represents a new move made (and thus a new set of features that need to
be added or subtracted). We also keep track of whether the values in the
accumulator are stale or not.

When we want to evaluate, we start at the most recent non-stale network
and ripple forwards, adding and subtracting as we go, until all
accumulators in the stack reflect the true values of the accumulators
for the board position after making those moves.

This is a huge speedup:

Elo   | 33.13 +- 10.96 (95%)
SPRT  | 8.0+0.08s Threads=1 Hash=16MB
LLR   | 2.96 (-2.94, 2.94) [0.00, 5.00]
Games | N: 1378 W: 408 L: 277 D: 693
Penta | [14, 106, 342, 189, 38]