Improve UciOption implementation
The existing implementation required a huge amount of duplication for
different options, which all needed to implement roughly similar parsing
code and needed to be manually inserted in both the set and print option
code paths.

Here, we instead keep an array of Box<dyn UciOption>, with callbacks for
when the options are set. These callbacks have a bit less flexibility as
they need to be given access to specific fields, but that likely
shouldn't be an issue.