Fix en-passant discovered check bug with files
There is already code to prevent generating moves that would give
discovered check on ranks - i.e. when our king and an enemy's rook or
queen are on the same rank, and an en-passant capture removes both pawns
blocking the check.

To check this, we remove both pawns from the board and check that
there's no new check. If there is a new check, we don't generate the
move.

However, we didn't place the capturing pawn back on the board (because
that wouldn't be relevant for a check on a rank). If we were instead in
a situation where our king and an enemy slider were on the same file,
this means we'd see a discovered check when - in reality - the capturing
pawn would block it.

Placing the pawn on the board where the move is played out prevents
this.

Thanks to Werner for Talkchess for the report:
https://talkchess.com/viewtopic.php?p=986038

While performing SPRT regression testing, the baseline engine crashed
when this situation was encountered, since the fixed engine would make
the move, but the original would consider it illegal. This happened 23
times out of 30954 games, so 0.07% of the time, or one in every ~1354
games.

Elo   | -0.09 +- 2.24 (95%)
SPRT  | 8.0+0.08s Threads=1 Hash=16MB
LLR   | 3.01 (-2.94, 2.94) [-5.00, 0.00]
Games | N: 30954 W: 7970 L: 7978 D: 15006
Penta | [413, 3542, 7554, 3576, 392]