Make the transposition table thread safe
This is broken out from the SMP implementation. This allows the
transposition table to be shared between multiple threads, sharing work
with eachother via the sharing of cache entries.

The bulk of the change is the creation of two new structs:

* TranspositionTableEntryBits, which represents the bits that make up a
  TranspositionTableEntry, free of their interpretation. One entry is 16
  bytes, so can be represented with two u64s.

* AtomicTranspositionTableEntry, which is what actually makes up the
  entries of the transposition table. This holds two AtomicU64 entries,
  which store one of the u64s in TranspositionTableEntryBits
  respectively.

We can only read/write one of these u64s atomically, so we're subject to
the possibility of races where we read two halves of different entries.
To protect ourselves, we use a technique from Robert Hyatt and Tim Mann,
XORing one u64 into the other. Since the first u64 represents the key of
the position, the key can only be recovered intact if XORed back with
the correct data for that position. If we do retrieve mismatched
entries, the key won't match the position and we'll consider it a miss.

We also take the opportunity to fix a couple of small issues:

* Age was truncated via storage in BoundAndAge, but not similarly
  truncated in the TT causing these to become mismatched. This didn't
  matter, but will now because:

* We now correctly report hashfull as entries for the current position,
  rather than non-empty slots. Reporting non-empty would cause hashfull
  to essentially be stuck at 100% shortly into games.

Elo   | 3.08 +- 6.48 (95%)
SPRT  | 8.0+0.08s Threads=1 Hash=16MB
LLR   | 3.10 (-2.94, 2.94) [-10.00, 0.00]
Games | N: 3604 W: 975 L: 943 D: 1686
Penta | [45, 390, 902, 418, 47]