Avoid borrowing persistent_state in SearchContext
This change is broken out from SMP to get most of the syntactic noise
out of the way.

Breaking things out when constructing SearchContext makes it easier to
pass a copy of HistoryTable in, instead of the 'main' table.

Also includes some very minor cleanups for the code that deals with TT
hits at the root.