Start storing NNUE networks out of the repository
This is pretty standard practice, since the individual files are
megabytes in size, and right now are changing fairly frequently. This
could balloon the repo size pretty quickly.

As with many other engines, the networks are stored in a release on
Github, and are downloaded while building if they're not already
present.

This does tie us to Github somewhat, and means if it were to go down,
past builds will fail, but on balance the tradeoff is worth it.

Also adjusted the naming scheme so that:
* The version number is included. This is generally nicer but also means
  that they're sorted correctly in a list.

* Removed -network and suffixed with .nnue instead since -network
  becomes redundant in this case.

* Used 8 chars in the SHA instead of 5, since this is what OpenBench
  uses so it's easier to compare visually.