Avoid accidentally un-setting the stop flag
After the previous fix for time losses during threefold repetitions, we
were still seeing occasional time losses.

An example of the issue can be seen by executing the following UCI
commands:

    position fen 5rk1/p2R1p1p/8/2P2p2/5B2/6Pn/P3r2P/R6K b - - 2 27
    go wtime 2000 btime 2000
    move Rc8 Rc1
    go wtime 200 btime 200 winc 200 binc 200

For the second search, the time strategy is as follows:

* Soft stop: 79.5ms
* Hard stop: 100ms

But the time taken is 193ms, almost double the hard stop.

The reason is the behaviour of TimeStrategy::update, which
unconditionally sets the stop flag to the result of .should_stop(). As a
result, we can have the following scenario:

* .update() is run in a negamax iteration and the stop flag is set. Our
  nodes will be a multiple of 10000, since that's how often we check.
  For this example, let's say it's 10000. Our next_check_at flag is
  incremented by 10000, meaning we're supposed to check again at 20000
  nodes.

* Immediately afterwards, we check the stop flag and return, since it
  was set.

* In the calling negamax, we progress until the next negamax call. For
  example, we might have just done our first PVS search and will drop
  into another. We drop down into negamax.

* In the new negamax, we immediately increment nodes_visited and then
  call .update(). nodes_visited is now 10001 and next_check_at is 20000.
  We unconditionally set the stop flag to the result of should_stop.
  Since nodes_visited < self.next_check_at, ths **returns false** and we
  un-set the stop flag.

This can happen multiple times, meaning we far exceed our desired hard
bound on the search time.

Also inline should_stop into update which prevents needing to do two
checks every update() call as most of the time we'll branch on the node
check.