Add tool to measure eval statistics
Based heavily on Cosmo's tool for measuring eval stats on a book.

Switched back to the very first NNUE network to get a sense of how scale
has changed over time.

The (abs) average on the original network was 367.27. The average of the
current v9 net is 852.58!

So to match the scale of the original net, we'd need to scale down the
output of v9 by ~172.30/400 = ~0.43.