nnue: Reduce scaling factor to match network #1
Over time as the proportion of WDL used during training has increased,
the output of the network has become inflated.

Measuring using tools/nnue_scaling, we can see that the average outputs
of the network over the lichess-big3 books are:

    Network  1: 367.27
    Network 10: 852.58

As a result, the engine's reported evalution differs from its original
eval by almost 2x. This can be seen in unbalanced positions during
online matches, where the engine often reports ~2x the other engine's
evaluation.

To get back to 'normal', we can adjust the scaling factor so that the
output of the network is scaled further.

Full stats:

Unscaled:

    Average: -246.52
    Average (abs): 838.36
    Max: 12342
    Min: -13334

Scaled:

    Average: -105.96
    Average (abs): 360.21
    Min: -5733
    Max: 5307

This has the side effect of altering the effect of various parameters
in search, making it more/less likely to do various kinds of pruning. In
order to not regress, we also SPSA-tuned these parameters.