Add a search stack
We've done without this until now by storing most information in
Game::history, but it doesn't really make sense to store things like
excluded moves or evals there as they're not used to undo moves, only
for search.

We can move our banning of null move pruning to use this as a first use
case.